requirements.txt
tensorflow>=2.13.0
pyyaml>=6.0
ipython>=8.0.0
jupyter>=1.0.0
seaborn>=0.12.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
keras>=2.13.0

